server/storage/requirements.txt
boto3==1.9.149
botocore==1.12.149
docutils==0.14
ibm-cos-sdk-core==2.4.4
ibm-cos-sdk-s3transfer==2.4.4
jmespath==0.9.4
paho-mqtt==1.4.0
python-dateutil==2.8.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.3
